Moved Setup to SetupService Moved the setup methods to a setup service class to reduce clutter of the backend controller class. Additionally, added response parsing on the front end to determine what information to be appended to the output. This should probably be refactored according to the notes mentioned.
